crawler_reclameaqui/ETL/requirements.txt
alembic==1.5.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cycler==0.10.0
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
nltk==3.5
numpy==1.20.0
pandas==1.2.0
Pillow==8.1.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
regex==2020.11.13
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
tqdm==4.56.2
urllib3==1.26.4
Werkzeug==1.0.1
wordcloud==1.8.1
